UI Changes And Knowledge Map
