+++ frontend/src/components/constant/constant.js
@@ -117,6 +117,14 @@ export const ReferenceViewScreen = {
     "Download Sample_Customer.csv"
   ]
 }
+
+
+export const KnowledgeMapScreen = {
+  docNames: ["Release Notes", "FSB", "Tech Notes", "Test Plan", "MOP"],
+  breadcrumb: "Knowledge Map",
+  tableHeader: "Knowledge Map Information :",
+}
+
 export const SolutionScreen = {
   breadcrumb: "Solution",
   modalHeader: "Patches",
@@ -130,7 +138,7 @@ export const SolutionScreen = {
   logInfo: "Log Info",
   tags: "Tags",
   filterButtons: ["All", "Clear"],
-  filterNames: ["Product", "Group", " Severity", "Priority", "Found on Platform", "Fixed in Release", "Found in Release", "Date Closed"],
+  filterNames: ["Product", "Group", " Severity", "Priority", "Found on Platform", "Fixed in Release", "Found in Release", "Date Closed On or Before"],
   tarBall: "TAR Ball",
   labAvailablity: "Lab Availablity",
   buttons: ["Predict", "Reserve", "Refine"],
