+++ frontend/src/router/index.js
@@ -18,6 +18,7 @@ import ChangePassword from '@/components/ManageRoles/ChangePassword'
 import ManageUser from '@/components/ManageRoles/ManageUser'
 import SolutionScreen from '@/components/MockUp/SolutionScreen'
 import MockUpNew from '@/components/MockUp/MockUpNew'
+import TestPlanKnowledgeMap from '@/components/KnowledgeMap/TestPlanKnowledgeMap'
 import KnowledgeMap from '@/components/KnowledgeMap/KnowledgeMap'
 
 Vue.use(Router)
@@ -152,6 +153,14 @@ export default new Router({
       component: MockUpNew,
       meta: {
         permission: 'Dashboard'
+      }
+    },
+    {
+      path: '/knowledge/testplan',
+      name: 'KnowledgeMap',
+      component: TestPlanKnowledgeMap,
+      meta: {
+        permission: 'knowledgeMap'
       }
     },
     {
